+++ spec/features/early_career_payments_claim_spec.rb
@@ -21,6 +21,12 @@ RSpec.feature "Teacher Early Career Payments claims" do
     expect(eligibility.nqt_in_academic_year_after_itt).to eql true
 
     # TODO [PAGE 02] - Which school do you teach at
+    expect(page).to have_text(I18n.t("early_career_payments.school_search"))
+
+    #choose_school schools(:penistone_grammar_school)
+    #expect(claim.eligibility.reload.school_search).to eql schools(:penistone_grammar_school)
+    
+    
     # TODO [PAGE 03] - Select the school you teach at
     # [PAGE 04] - Are you currently employed as a supply teacher
     expect(page).to have_text(I18n.t("early_career_payments.questions.employed_as_supply_teacher"))