+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog]
     - Add 'Are you currently subject to disciplinary action?'
     - Add 'Did you do a postgraduate ITT course or an undergraduate ITT course?'
     - Add 'You are not eligible for an early careers payment'
+    - Add 'Which school do you teach at?'
 - Automatically update admin claim tasks with DQT API
 
 ## [Release 089] - 2021-03-05